Revert "HomeWork Exception done"
This reverts commit ffff57c3eeaf6c5387db2e79e2fd170c43870adf.
