--- ExceptionUsage_ HW/ExceptionUsage_ HW.sln
@@ -1,22 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.12.35707.178 d17.12
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExceptionUsage_ HW", "ExceptionUsage_ HW.csproj", "{53B1559C-C3CA-49D0-A477-6EE706E4BDE4}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{53B1559C-C3CA-49D0-A477-6EE706E4BDE4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{53B1559C-C3CA-49D0-A477-6EE706E4BDE4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{53B1559C-C3CA-49D0-A477-6EE706E4BDE4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{53B1559C-C3CA-49D0-A477-6EE706E4BDE4}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
